move logparser and logpipeline validation wekhooks to the new versioned package
