+++ .golangci.yaml
@@ -66,10 +66,8 @@ linters-settings:
         alias: kitk8s
       - pkg: github.com/kyma-project/telemetry-manager/test/testkit/kyma
         alias: kitkyma
-      - pkg: github.com/kyma-project/telemetry-manager/webhook/logparser
-        alias: logparserwebhook
-      - pkg: github.com/kyma-project/telemetry-manager/webhook/logpipeline
-        alias: logpipelinewebhook
+      - pkg: github.com/kyma-project/telemetry-manager/webhook/logparser/v1alpha1
+        alias: logparserwebhookv1alpha1
       - pkg: github.com/kyma-project/telemetry-manager/webhook/logpipeline/v1alpha1
         alias: logpipelinewebhookv1alpha1
       - pkg: github.com/kyma-project/telemetry-manager/webhook/logpipeline/v1beta1
