+++ webhook/logparser/v1alpha1/spec_validator.go
@@ -1,4 +1,4 @@
-package logparser
+package v1alpha1
 
 import (
 	"fmt"
