+++ webhook/logpipeline/v1alpha1/files_validator.go
@@ -1,4 +1,4 @@
-package logpipeline
+package v1alpha1
 
 import (
 	"fmt"
