+++ webhook/logpipeline/v1alpha1/spec_validator_test.go
@@ -1,4 +1,4 @@
-package logpipeline
+package v1alpha1
 
 import (
 	"context"
