+++ webhook/logpipeline/v1alpha1/variable_validator.go
@@ -1,4 +1,4 @@
-package logpipeline
+package v1alpha1
 
 import (
 	"fmt"
