+++ webhook/logpipeline/v1alpha1/variable_validator_test.go
@@ -1,4 +1,4 @@
-package logpipeline
+package v1alpha1
 
 import (
 	"context"
